Add option to pass the file mode
This change allows the user to specify the changes as an object
that has the path of the file as a key and an object as value.

The object will contain the content key that represent the content
of the file you want to modify, if null it will delete the file.
And it can contain the mode, if no mode is specified the library falls
back to the previous behaviour assigning the default 100644 mode